requirements.txt
selenium
opencv-python
pytesseract
pillow
numpy
openai
pyyaml
pyautogui
pytest
python-dotenv
requests
webdriver-manager
loguru
retry 